src/main/java/net/chesstango/board/movesgenerators/pseudo/strategies/KnightMoveGenerator.java
package net.chesstango.board.movesgenerators.pseudo.strategies;

import net.chesstango.board.Color;
import net.chesstango.board.PiecePositioned;
import net.chesstango.board.Square;
import net.chesstango.board.iterators.bysquare.bypiece.KnightPositionsSquareIterator;
import net.chesstango.board.moves.Move;

import java.util.Iterator;


/**
 * @author Mauricio Coria
 *
 */
public class KnightMoveGenerator extends AbstractJumpMoveGenerator {
	
	public KnightMoveGenerator(Color color) {
		super(color);
	}
	
	@Override
	protected Move createSimpleMove(PiecePositioned origen, PiecePositioned destino) {
		return this.moveFactory.createSimpleMove(origen, destino, null);
	}


	@Override
	protected Move createCaptureMove(PiecePositioned origen, PiecePositioned destino) {
		return this.moveFactory.createCaptureMove(origen, destino, null);
	}

	@Override
	protected Iterator<Square> getSquareIterator(Square fromSquare) {
		return new KnightPositionsSquareIterator(fromSquare);
	}

}
